moved telemetryData test to separate file
