+++ tests/packet/test_packet.py
@@ -285,11 +285,3 @@ def test_packet_print_no_body(packet):
                       "No body"
 
     assert expected_string == packet.__str__()
-
-
-def test_telemetry_data():
-    current_time = datetime.now().timestamp()
-    new_data = TelemetryData(current_time, 0, 0, 0, 0, 0, 0)
-    valid = new_data.set_validity()
-    encoded_new_data = new_data.encode_data()
-    decoded_new_data = TelemetryData.decode_data(encoded_new_data)
